Chart is displaying on homepage
The database is queried with either class_name or instructor_name and the grade distribution is displayed.
There is a component called Graph which holds the chart generation function, so this can be easily added to other components.

I still have to make it so that it accepts specific querying arguments instead of requesting example data.